small normings
