+++ Makefile
@@ -6,7 +6,7 @@
 #    By: kwurster <kwurster@student.42berlin.de>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/04/22 15:55:27 by lmeubrin          #+#    #+#              #
-#    Updated: 2025/04/23 14:24:32 by lmeubrin         ###   ########.fr        #
+#    Updated: 2025/04/28 12:34:16 by lmeubrin         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -49,7 +49,8 @@ OPTIM_FLAGS := -O3 -march=native
 LINUX_MLX := -ldl -lglfw -pthread -lm
 MAC_MLX := -lglfw -framework Cocoa -framework OpenGL -framework IOKit
 
-FINAL_CFLAGS = $(CFLAGS) $(OPTIM_FLAGS)
+#TODO: IMPORTANT, clear the bonus flag when submitting mandatory
+FINAL_CFLAGS = $(CFLAGS) $(OPTIM_FLAGS) -DBONUS
 FINAL_LDFLAGS = $(LDFLAGS)
 
 # Different sanitizer configurations
@@ -85,9 +86,11 @@ SRCS := $(addprefix $(SRCS_DIR)/,\
 		aabb.c \
 		window_utils.c \
 		scene_fixing.c \
-		$(addprefix util/, quaternion.c color.c util.c vec3_0.c vec3_1.c vec3_2.c scene_malloc.c \
-		print_scene.c print_scene_utils_0.c print_scene_utils_1.c interval.c preprocess.c preprocess2.c) \
-		$(addprefix parse/, input_check.c parse_scene.c set_scene.c set_objs.c set_utils.c) \
+		$(addprefix util/, quaternion.c color.c util.c vec3_0.c vec3_1.c vec3_2.c \
+		scene_malloc.c scene_destroy interval.c preprocess.c preprocess2.c \
+		print_scene.c print_scene_utils_0.c print_scene_utils_1.c) \
+		$(addprefix parse/, input_check.c parse_scene.c set_scene.c set_objs.c \
+		set_utils.c) \
 		)
 
 OBJS := $(SRCS:($SRCS_DIR)%.c=$(OBJ_DIR)/%.o)
