+++ srcs/util/util.c
@@ -6,7 +6,7 @@
 /*   By: kwurster <kwurster@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/18 13:55:31 by kwurster          #+#    #+#             */
-/*   Updated: 2025/04/25 12:34:07 by lmeubrin         ###   ########.fr       */
+/*   Updated: 2025/04/28 12:30:46 by lmeubrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,60 +16,6 @@
 float	clamp(float value, float min, float max)
 {
 	return (fminf(fmaxf(value, min), max));
-}
-
-void	objects_destroy(t_objects* objects, bool free_inner)
-{
-	size_t	i;
-
-	if (!objects)
-		return ;
-	i = 0;
-	if (free_inner)
-	{
-		while (i < objects->sphere_count)
-			free(objects->spheres[i++].lights);
-	}
-	free(objects->spheres);
-	objects->sphere_count = 0;
-	free(objects->planes);
-	objects->plane_count = 0;
-	i = 0;
-	if (free_inner)
-	{
-		while (i < objects->cylinder_count)
-			free(objects->cylinders[i++].lights);
-	}
-	free(objects->cylinders);
-	objects->cylinder_count = 0;
-	i = 0;
-	if (free_inner)
-	{
-		while (i < objects->cone_count)
-			free(objects->cones[i++].lights);
-	}
-	free(objects->cones);
-	objects->cone_count = 0;
-}
-
-/// Free memory allocated for a scene
-void	scene_destroy(t_scene *scene)
-{
-	size_t	i;
-
-	if (!scene)
-		return ;
-	i = 0;
-	while (i < scene->light_count)
-		objects_destroy(&scene->lights[i++].objs, false);
-	free(scene->lights);
-	if (scene->light_count == 0)
-		objects_destroy(&scene->objs, false);
-	else
-	{
-		scene->light_count = 0;
-		objects_destroy(&scene->objs, true);
-	}
 }
 
 float	image_aspect_ratio(t_scene *scene)
